:sparkles: POC for Results display
This is very rough, just showing the raw JSON more-or-less. Haven't polished the GQL query, and the variables are hardcoded in. Also, to work on this locally, you'll need a .env file with a REACT_APP_GH_KEY variable for hitting the GitHub API. I'll post instructions on how to get one set up when I can.