:truck: Rename query to search
There's some naming collisions going on here: we have "queries" in GraphQL, a "query string" as part of the URL, and what I've (up til now) called a "query string" that will be part of the GraphQL query (literally as the variable query). To make this hopefully less confusing, I'm renaming the last of those to "search string". So, moving forward "query string" will only refer to variables attached to the end of the URL.